+++ src/index.ts
@@ -1,4 +1,5 @@
 import Star from "./star";
 import Fork from "./fork";
+import Watch from "./watch";
 
-export { Fork, Star };
+export { Fork, Star, Watch };