+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All notable changes to the "react-github-buttons" project will be documented in this file.
 
+## [Unreleased]
+### Added
+
+- Watch Button
+
 ## [0.1.0] 2019-6-13
 ### Added